SSCS-3375 Check postcode is in England or Wales
Split out object for branch for England and Wales
