--- steps/start/postcode-checker/southEastPostcodes.js
@@ -1,636 +1,0 @@
-/* eslint-disable max-lines */
-const { concat } = require('lodash');
-
-const ashford = [
-  'CT14',
-  'CT15',
-  'CT16',
-  'CT17',
-  'CT18',
-  'CT19',
-  'CT20',
-  'CT21',
-  'ME7',
-  'ME8',
-  'ME9',
-  'ME10',
-  'ME11',
-  'ME12',
-  'ME13',
-  'ME14',
-  'ME15',
-  'ME16',
-  'ME17',
-  'ME18',
-  'ME19',
-  'ME20',
-  'TN1',
-  'TN2',
-  'TN3',
-  'TN4',
-  'TN5',
-  'TN9',
-  'TN10',
-  'TN11',
-  'TN12',
-  'TN17',
-  'TN18',
-  'TN19',
-  'TN20',
-  'TN21',
-  'TN23',
-  'TN24',
-  'TN25',
-  'TN26',
-  'TN27',
-  'TN28',
-  'TN29',
-  'TN30',
-  'TN31',
-  'TN32',
-  'TN33',
-  'TN34',
-  'TN35',
-  'TN36',
-  'TN37',
-  'TN38',
-  'TN39',
-  'TN40'
-];
-
-const basildon = [
-  'SS5',
-  'SS6',
-  'SS7',
-  'SS8',
-  'SS9',
-  'SS10',
-  'SS11',
-  'SS12',
-  'SS13',
-  'SS14',
-  'SS15',
-  'SS16',
-  'SS17',
-  'CM4',
-  'CM11',
-  'CM12',
-  'CM13',
-  'CM14',
-  'CM15'
-];
-
-const bedford = [
-  'MK40',
-  'MK41',
-  'MK42',
-  'MK43',
-  'MK44',
-  'MK45',
-  'MK46',
-  'SG15',
-  'SG16',
-  'SG17',
-  'SG18',
-  'SG19'
-];
-
-const brighton = [
-  'RH10',
-  'RH11',
-  'RH12',
-  'RH13',
-  'RH14',
-  'RH15',
-  'RH16',
-  'RH17',
-  'RH18',
-  'RH19',
-  'RH20',
-  'BN1',
-  'BN2',
-  'BN3',
-  'BN4',
-  'BN5',
-  'BN6',
-  'BN7',
-  'BN8',
-  'BN9',
-  'BN10',
-  'BN11',
-  'BN12',
-  'BN13',
-  'BN14',
-  'BN15',
-  'BN16',
-  'BN17',
-  'BN18',
-  'BN19',
-  'BN25',
-  'BN26',
-  'BN27',
-  'BN28',
-  'BN29',
-  'BN30',
-  'BN31',
-  'BN32',
-  'BN33',
-  'BN34',
-  'BN35',
-  'BN36',
-  'BN37',
-  'BN38',
-  'BN39',
-  'BN40',
-  'BN41',
-  'BN42',
-  'BN43',
-  'BN44',
-  'BN45',
-  'TN6',
-  'TN7'
-];
-
-const cambridge = [
-  'SG8',
-  'SG9',
-  'CB1',
-  'CB10',
-  'CB11',
-  'CB2',
-  'CB20',
-  'CB21',
-  'CB22',
-  'CB23',
-  'CB24',
-  'CB25',
-  'CB3',
-  'CB4',
-  'CB5',
-  'CB6',
-  'CB7',
-  'CB8',
-  'CB9',
-  'IP28',
-  'IP29',
-  'IP30',
-  'IP31',
-  'IP32',
-  'IP33',
-  'PE27',
-  'PE29',
-  'SG8',
-  'SG8'
-];
-
-const chatham = [
-  'ME1',
-  'ME2',
-  'ME3',
-  'ME4',
-  'ME5',
-  'ME6'
-];
-
-const colchester = [
-  'CM0',
-  'CM1',
-  'CM2',
-  'CM3',
-  'CM7',
-  'CM8',
-  'CM9',
-  'CM77',
-  'CO1',
-  'CO2',
-  'CO3',
-  'CO4',
-  'CO5',
-  'CO6',
-  'CO7',
-  'CO8',
-  'CO9'
-];
-
-const eastbourne = [
-  'TN41',
-  'BN20',
-  'BN21',
-  'BN22',
-  'BN23',
-  'BN24'
-];
-
-const harlow = [
-  'CM10',
-  'CM16',
-  'CM17',
-  'CM18',
-  'CM19',
-  'CM20',
-  'CM21',
-  'CM22',
-  'CM23',
-  'CM24',
-  'CM25',
-  'CM5',
-  'CM6'
-];
-
-const ipswich = [
-  'CO10',
-  'CO11',
-  'CO12',
-  'CO13',
-  'CO14',
-  'CO15',
-  'CO16',
-  'CO4',
-  'CO7',
-  'IP1',
-  'IP2',
-  'IP3',
-  'IP4',
-  'IP5',
-  'IP6',
-  'IP7',
-  'IP8',
-  'IP9',
-  'IP10',
-  'IP11',
-  'IP12',
-  'IP13',
-  'IP14',
-  'IP15',
-  'IP16',
-  'IP17'
-];
-
-const kingsLynn = [
-  'PE13',
-  'PE14',
-  'PE15',
-  'PE30',
-  'PE31',
-  'PE32',
-  'PE33',
-  'PE34',
-  'PE35',
-  'PE36',
-  'PE37',
-  'PE38'
-];
-
-const luton = [
-  'AL1',
-  'AL3',
-  'AL5',
-  'HP2',
-  'LU1',
-  'LU2',
-  'LU3',
-  'LU4',
-  'LU5',
-  'LU6',
-  'LU7'
-];
-
-const highWycombe = [
-  'SL0',
-  'SL1',
-  'SL2',
-  'SL3',
-  'SL4',
-  'SL5',
-  'SL6',
-  'SL7',
-  'SL8',
-  'SL9',
-  'UB10',
-  'UB11',
-  'UB3',
-  'UB4',
-  'UB7',
-  'UB8',
-  'UB9'
-];
-
-const miltonKeynes = [
-  'MK1',
-  'MK2',
-  'MK3',
-  'MK4',
-  'MK5',
-  'MK6',
-  'MK7',
-  'MK8',
-  'MK9',
-  'MK10',
-  'MK11',
-  'MK12',
-  'MK13',
-  'MK14',
-  'MK15',
-  'MK16',
-  'MK17',
-  'MK18',
-  'MK19',
-  'MK20',
-  'MK21',
-  'MK22',
-  'MK23',
-  'MK24',
-  'MK25',
-  'MK26',
-  'MK27',
-  'MK28',
-  'MK29',
-  'MK30',
-  'MK31',
-  'MK32',
-  'MK33',
-  'MK34',
-  'MK35',
-  'MK36',
-  'MK37',
-  'MK38',
-  'MK39'
-];
-
-const norwich = [
-  'IP18',
-  'IP19',
-  'IP20',
-  'IP21',
-  'IP22',
-  'IP23',
-  'IP24',
-  'IP25',
-  'IP26',
-  'IP27',
-  'NR1',
-  'NR2',
-  'NR3',
-  'NR4',
-  'NR5',
-  'NR6',
-  'NR7',
-  'NR8',
-  'NR9',
-  'NR10',
-  'NR11',
-  'NR12',
-  'NR13',
-  'NR14',
-  'NR15',
-  'NR16',
-  'NR17',
-  'NR18',
-  'NR19',
-  'NR20',
-  'NR21',
-  'NR22',
-  'NR23',
-  'NR24',
-  'NR25',
-  'NR26',
-  'NR27',
-  'NR28',
-  'NR29',
-  'NR30',
-  'NR31',
-  'NR32',
-  'NR33',
-  'NR34',
-  'NR35'
-];
-
-const oxford = [
-  'OX1',
-  'OX2',
-  'OX3',
-  'OX4',
-  'OX5',
-  'OX6',
-  'OX7',
-  'OX8',
-  'OX9',
-  'OX10',
-  'OX11',
-  'OX12',
-  'OX13',
-  'OX14',
-  'OX15',
-  'OX16',
-  'OX17',
-  'OX18',
-  'OX19',
-  'OX20',
-  'OX21',
-  'OX22',
-  'OX23',
-  'OX24',
-  'OX25',
-  'OX26',
-  'OX27',
-  'OX28',
-  'OX29',
-  'OX30',
-  'OX31',
-  'OX32',
-  'OX33',
-  'OX34',
-  'OX35',
-  'OX36',
-  'OX37',
-  'OX38',
-  'OX39',
-  'OX40',
-  'OX41',
-  'OX42',
-  'OX43',
-  'OX44',
-  'OX45',
-  'OX46',
-  'OX47',
-  'OX48',
-  'OX49'
-];
-
-const peterborough = [
-  'PE1',
-  'PE2',
-  'PE3',
-  'PE4',
-  'PE5',
-  'PE6',
-  'PE7',
-  'PE8',
-  'PE9',
-  'PE10',
-  'PE16',
-  'PE17',
-  'PE18',
-  'PE19',
-  'PE26',
-  'PE28'
-];
-
-const reading = [
-  'RG1',
-  'RG2',
-  'RG3',
-  'RG4',
-  'RG5',
-  'RG6',
-  'RG7',
-  'RG8',
-  'RG9',
-  'RG10',
-  'RG11',
-  'RG12',
-  'RG13',
-  'RG14',
-  'RG15',
-  'RG16',
-  'RG17',
-  'RG18',
-  'RG19',
-  'RG20',
-  'RG21',
-  'RG22',
-  'RG23',
-  'RG24',
-  'RG25',
-  'RG26',
-  'RG27',
-  'RG28',
-  'RG29',
-  'RG30',
-  'RG31',
-  'RG32',
-  'RG33',
-  'RG34',
-  'RG35',
-  'RG36',
-  'RG37',
-  'RG38',
-  'RG39',
-  'RG40',
-  'RG41',
-  'RG42',
-  'RG43',
-  'RG44',
-  'RG45'
-];
-
-const southend = [
-  'SS0',
-  'SS1',
-  'SS2',
-  'SS3',
-  'SS4'
-];
-const stevenage = [
-  'SG1',
-  'SG2',
-  'SG3',
-  'SG4',
-  'SG5',
-  'SG6',
-  'SG7',
-  'SG10',
-  'SG11',
-  'SG12',
-  'SG13',
-  'SG14'
-];
-
-const watford = [
-  'AL2',
-  'AL4',
-  'HA1',
-  'HA2',
-  'HA3',
-  'HA4',
-  'HA5',
-  'HA6',
-  'HA7',
-  'HP1',
-  'HP3',
-  'HP4',
-  'HP5',
-  'HP6',
-  'HP7',
-  'HP8',
-  'HP9',
-  'HP10',
-  'HP11',
-  'HP12',
-  'HP13',
-  'HP14',
-  'HP15',
-  'HP16',
-  'HP17',
-  'HP18',
-  'HP19',
-  'HP20',
-  'HP21',
-  'HP22',
-  'HP23',
-  'HP24',
-  'HP25',
-  'HP26',
-  'HP27',
-  'UB5',
-  'UB6',
-  'WD1',
-  'WD2',
-  'WD3',
-  'WD4',
-  'WD5',
-  'WD6',
-  'WD7',
-  'WD8',
-  'WD9',
-  'WD10',
-  'WD11',
-  'WD12',
-  'WD13',
-  'WD14',
-  'WD15',
-  'WD16',
-  'WD17',
-  'WD18',
-  'WD19',
-  'WD20',
-  'WD21',
-  'WD22',
-  'WD23',
-  'WD24',
-  'WD25'
-];
-
-module.exports = concat(
-  ashford,
-  basildon,
-  bedford,
-  brighton,
-  cambridge,
-  chatham,
-  colchester,
-  eastbourne,
-  harlow,
-  ipswich,
-  kingsLynn,
-  luton,
-  highWycombe,
-  miltonKeynes,
-  norwich,
-  oxford,
-  peterborough,
-  reading,
-  southend,
-  stevenage,
-  watford
-);
